lab3

+++ lab3/Account.java
@@ -5,81 +5,79 @@ abstract class Account{
 
 
     /**
-     * This method will set the account number and balance of the user account.
+     * constructor
      * @param accNum
      * @param b
      * @return
      */
-    // constructor
     Account(int accNum, double b){
         this.accountNumber = accNum;
         this.startingBalance = b;
         this.balance = b;
-        //this.logIndex = 0;
     }
 
+    /**
+     * getStartingBalance: gets current user starting account balance
+     * @return double: account balance
+     */
     public double getStartingBalance(){
       return this.startingBalance;
     }
     /**
-     * This method will allow us the update the private attribute of balance
-     * @param amount
+     * updateBalance: This method will allow us the update the private attribute of balance
+     * @param amount double: new account balance
      */
     // setters
     public void updateBalance(double amount){
         this.balance = amount;
     }
 
-
     /**
-     * This method will allow you to retrieve the private
+     * getAccountNumber: This method will allow you to retrieve the private
      * attribute account number
-     * @return int
+     * @return int: account number
      */
     // getters
     public int getAccountNumber(){
         return this.accountNumber;
     }
 
-
     /**
-     * This method will allow you to retrieve the private balance
-     * @return double
+     * getBalance: This method will allow you to retrieve the private balance
+     * @return double: retuns current balance
      */
     public double getBalance(){
         return this.balance;
     }
-
-
-
     /**
-     * Abstract withdraw method. Since not all accounts will be able
-     * to withdraw (credit). This method can be abstract and override
-     * it to no do nothing in certain circumstances
-     * @param transferAccount
-     * @param amount
-     * @return boolean
+     * withdraw: abstract withdraw method, allows customer to witdraw from their
+     * accounts
+     * @param  amount double: how much to withdraw
+     * @return        boolean: succesful
      */
-    // functions
     public abstract boolean withdraw(double amount);
 
-    public abstract int getMax();
+    /**
+     * getMax: gets the credit max
+     * @return double: credit max value
+     */
+    public abstract double getMax();
 
     /**
-     * Abstract transfer method. Since not all accounts will be able
+     * transfer: Abstract transfer method. Since not all accounts will be able
      * to transfer (credit). This method can be abstract and override
      * it to no do nothing in certain circumstances
-     * @param transferAccount
-     * @param amount
-     * @return boolean
+     * @param transferAccount Account: other user account
+     * @param amount double: how much to transfer
+     * @return boolean: succesful
      */
     public abstract boolean transfer(Account transferAccount, double amount);
 
 
     /**
-     * This method will allow the user to make a deposit to all their accounts
-     * @param amount
-     * @return boolean
+     * deposit: This method will allow the user to make a deposit to all their accounts
+     * @param amount double: how much to deposit
+     * @return boolean: succesful
      */
     public boolean deposit(double amount){
         // checks if amount is negative
